(refactor) Ensure database code is in sync with build system changes
