+++ package.json
@@ -19,6 +19,7 @@
   "homepage": "https://github.com/relentlessbreakfast/constellation",
   "dependencies": {
     "bluebird": "^2.9.27",
+    "body-parser": "^1.12.4",
     "fs": "0.0.2",
     "github": "^0.2.4",
     "morgan": "^1.5.3",
